Move entry point of app to the main.rb
